Update: Remove Comment in some lines
To make the code easier to test out and experiment on, I 'uncommented' some of the examples so that they execute when the code is run in the terminal. 

You can always comment the unnecessary lines by adding a "#" at the beginning of the line. Example:

print("Hello World") <=== Uncomment
# print("Hello World") <=== Commented i.e. this code will not run.